+++ AirportBrcmFixup/kern_config.hpp
@@ -9,6 +9,7 @@
 #define kern_config_private_h
 
 #include <Headers/kern_util.hpp>
+#include <IOKit/IOService.h>
 
 
 class Configuration {
@@ -16,32 +17,36 @@ public:
 	/**
 	 *  Possible boot arguments
 	 */
-	static const char *bootargOff[];
-	static const char *bootargDebug[];
-	static const char *bootargBeta[];
-	static constexpr const char *bootargBrcmCountry {"brcmfx-country"};     // 5 Ghz patch - change default country
-	static constexpr const char *bootargBrcmEnableWowl {"brcmfx-wowl"};     // enable WOWL
-	static constexpr const char *bootargBrcmDriver {"brcmfx-driver"};
-	static constexpr const char *bootargBrcmAspm {"brcmfx-aspm"};
-
+	static constexpr const char *bootargOff            {"-brcmfxoff"};
+	static constexpr const char *bootargDebug          {"-brcmfxdbg"};
+	static constexpr const char *bootargBeta           {"-brcmfxbeta"};
+	static constexpr const char *bootargBrcmEnableWowl {"-brcmfx-wowl"};        // enable WOWL
+	static constexpr const char *bootargBrcmAllDrv     {"-brcmfx-alldrv"};	   // skip driver availability check (which drivers available in the current system)
+	static constexpr const char *bootargBrcmCountry    {"brcmfx-country"};     // 5 Ghz patch - change default country
+	static constexpr const char *bootargBrcmDriver     {"brcmfx-driver"};
+	static constexpr const char *bootargBrcmAspm       {"brcmfx-aspm"};
 
 	/**
 	 *  Retrieve boot arguments
 	 *
 	 *  @return true if allowed to continue
 	 */
-	void readArguments();
+	void readArguments(IOService* provider = nullptr);
+	bool awaitPublishing(IORegistryEntry *obj);
 
-	char country_code[4] {"US"};
-	bool country_code_overrided {false};
+	char country_code[5]        {"US"};
 	
-	bool disabled            {false};
-	bool enable_wowl         {false};
+	bool disabled               {false};
+	bool enable_wowl            {false};
 
-	int32_t brcmfx_driver    {-1};
+	int32_t brcmfx_driver       {-1};
 
-	IOOptionBits brcmfx_aspm {0};
-	bool override_aspm		 {false};
+	IOOptionBits brcmfx_aspm    {0};
+	bool override_aspm          {false};
+	
+	bool enable_all_drv         {false};
+	
+	bool config_is_ready        {false};
 
 	Configuration() = default;
 };