+++ AirportBrcmFixup/kern_fakebrcm.hpp
@@ -25,7 +25,6 @@ class EXPORT PCIHookManager
 
 public:
 	static void hookProvider(IOService* provider);
-	static bool awaitPublishing(IORegistryEntry *obj);
 };
 
 class EXPORT FakeBrcm : public IOService